Final touch for chat
Added the last things based of test-backend
